debian: mask alsa-restore
Our audio driver often sets some temporary state as mixer parameters.
If we save/restore them, we run the risk of leaving the audio stack in
an inconsistent state, which can cause a bunch of different issues, such
as microphones not working, wired headphones being noisy/leaky, etc.

We really don't need to persist any of these variables in the first place,
and doing so is a bit of an anachronism in the times of pulseaudio, g-s-d,
etc. So let's just not do that.